requirements.txt
awsebcli==3.18.2
beautifulsoup4==4.9.0
botocore==1.15.49
Brotli==1.0.9
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.3
dash==1.15.0
dash-auth==1.4.1
dash-bootstrap-components==0.10.5
dash-core-components==1.11.0
dash-html-components==1.1.0
dash-renderer==1.7.0
dash-table==4.10.0
docutils==0.15.2
Flask==1.1.2
Flask-Compress==1.5.0
Flask-SeaSurf==0.2.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.1
pathspec==0.5.9
pillow==6.0.0
plotly==4.9.0
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
retrying==1.3.3
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
ua-parser==0.10.0
urllib3==1.24.3
wcwidth==0.1.9
Werkzeug==1.0.1
